requirements.txt
rich
stanza
zhipuai
python-dotenv
loguru
tqdm
